rejigger on tool logging
